requirements.txt
python-dotenv
PyPDF2
tiktoken
openai>=1.0.0
